feat: add bundle visualizer and enhance contract CI
- Add rollup-plugin-visualizer for bundle analysis
- Create npm run build:analyze script for visualization
- Implement React Router lazy-loading for page routes
- Extract dashboard logic to separate DashboardPage component
- Split SenderDashboard and RecipientDashboard into lazy-loaded chunks
- Create frontend/PERFORMANCE.md with bundle optimization guide
- Fix contract-ci.yml syntax error (YAML indentation)
- Add cargo clippy with -D warnings flag to contract CI
- Add WASM binary size reporting to workflow summary
- Improve frontend performance with code-splitting

Closes #407 #413
