+++ frontend/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
+    "build:analyze": "tsc && vite build --mode analyze",
     "preview": "vite preview",
     "test": "vitest",
     "test:e2e": "npx playwright test --config=playwright.config.ts",
@@ -18,6 +19,7 @@
     "lucide-react": "^0.294.0",
     "react": "^18.2.0",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.0",
     "recharts": "^3.7.0",
     "zod": "^4.3.6"
   },
@@ -40,6 +42,7 @@
     "jsdom": "^29.0.1",
     "msw": "^2.12.14",
     "postcss": "^8.4.32",
+    "rollup-plugin-visualizer": "^5.12.0",
     "tailwindcss": "^3.4.0",
     "typescript": "^5.2.2",
     "typescript-eslint": "^8.60.1",
